fixed removing ships

+++ src/Controller/api/ShipPersonalController.php
@@ -8,17 +8,23 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
 use App\Entity\ShipPersonal;
 use Doctrine\ORM\EntityManagerInterface;
+use Psr\Log\LoggerInterface;
 
 class ShipPersonalController extends AbstractController
 {
     // delete the ship with priority, name, and current user
     #[Route('/api/shippersonal/delete', name: 'api_shippersonal_delete')]
-    public function deleteShip(EntityManagerInterface $em, Request $request): Response
+    public function deleteShip(EntityManagerInterface $em, Request $request, LoggerInterface $logger): Response
     {
         $email = $this->getUser()->getUserIdentifier();
         $name = $request->query->get('name');
         $priority = $request->query->get('priority');
         $shipPersonal = $em->getRepository(ShipPersonal::class)->findPrecise($email, $name, $priority);
+
+        if ($shipPersonal == null) {
+            $logger->warning("Can't find shipPersonal with " . $name . " of " . $email . " in ShipPersonalController /delete");
+            return $this->redirectToRoute('app_home');
+        }
 
         $em->remove($shipPersonal);
         $em->flush();
